VFS: lift d_alloc_parallel above inode_lock
d_alloc_parallel() can block waiting on a d_in_lookup() dentry
so it is important to order it consistently with other blocking locks
such as inode_lock().

Currenty d_alloc_parallel() is ordered after inode_lock(): it can be
called while the inode is locked, and so the inode cannot be locked
while a d_in_lookup() dentry is held.

This patch reverses that order.  d_alloc_parallel() must now be called
*before* locking the directory, and must not be called afterwards.  This
allows directory locking to be moved closer to the filesystem
operations, and ultimately into those operations.

lookup_one_qstr_excl() is now called without an lock held, exclusive or
otherwise, so the "_excl" is dropped - it is now lookup_one_qstr().

As a lock is taken *after* lookup, start_dirop() and start_renaming()
must ensure that if the dentry isn't in-lookup then after the lock has
been taken the parent is still correct, the name is still correct, and
the dentry is still hashed.  New function dentry_matches() implements
this.  It takes a seq number from rename_lock so that the name
comparison can often be skipped.

The start_ function also need to recheck a dentry is negative if
LOOKUP_EXCL was in force.

lookup_slow() doesn't need to re-check the parent as the dentry is
always in-lookup so parent cannot change.

lookup_open() only needs to recheck ->d_inode before calling ->create().
If ->d_inode is still NULL, then the dentry is negative and cannot have
been renamed or unlinked.

The locking in lookup_slow() is moved into __lookup_slow() immediately
before/after ->lookup, and lookup_slow() just sets the task state for
waiting.

Parent locking is removed from open_last_lookups() and performed in
lookup_open().  A shared lock is taken if ->lookup() needs to be called.
An exclusive lock is taken separately if ->create() needs to be called -
with checks that the dentry hasn't become positive.

If ->atomic_open is needed we take exclusive or shared parent lock as
appropriate and check for a positive dentry or DEAD parent.

The fsnotify_create() call is kept inside the locked region in
lookup_open().  I don't know if this is important.

Signed-off-by: NeilBrown <neil@brown.name>