VFS: use new dentry locking for open/create/remove/rename
This patch updates start_dirop (used for create via start_creating and
remove via start_removing), start_renaming (used for rename) and
lookup_open/atomic_open (used for open), and related function.

In each case we now use dentry_lock in place of locking the parent, and
then check the dentry is still as expected.  The locking of the parent
is preserved but happens later - it is no longer needed to ensure the
dentry does not get moved, unhashed, or instantiated.  It is only needed
because some filesystems might depend on it.

As we no longer depend on the parent inode lock to block renames, we
don't need to take that lock in __d_unalias().

Signed-off-by: NeilBrown <neil@brown.name>